+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/base/irs/model/request/RegisterPolicyRequest.java
@@ -19,30 +19,13 @@
 
 package org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.request;
 
-import org.eclipse.tractusx.irs.edc.client.policy.Constraint;
-import org.eclipse.tractusx.irs.edc.client.policy.Constraints;
-import org.eclipse.tractusx.irs.edc.client.policy.OperatorType;
-import org.eclipse.tractusx.irs.edc.client.policy.Permission;
-import org.eclipse.tractusx.irs.edc.client.policy.PolicyType;
+import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.Payload;
 
 import java.time.Instant;
-import java.time.OffsetDateTime;
-import java.util.List;
 
 public record RegisterPolicyRequest(
-        String policyId,
         Instant validUntil,
-        List<Permission> permissions
+        Payload payload
 ) {
-    public static RegisterPolicyRequest from(String leftOperand, OperatorType operatorType, String rightOperand, OffsetDateTime ttl) {
-        return new RegisterPolicyRequest(
-                rightOperand,
-                ttl.toInstant(),
-                List.of(new Permission(
-                        PolicyType.USE,
-                        List.of(new Constraints(
-                                List.of(new Constraint(leftOperand, operatorType, List.of(rightOperand))),
-                                List.of(new Constraint(leftOperand, operatorType, List.of(rightOperand)))))
-                )));
-    }
+
 }